バグ修正
ロールオーバーの点灯処理とリプレイ処理のバグを修正
